requirements.txt
playwright>=1.40
requests>=2.31
google-api-python-client>=2.100
google-auth-httplib2>=0.1
google-auth>=2.23
jinja2>=3.1
